Updated files required to compile and run the GO implementaion of the node. (WIP)
